+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Add `cargo zng l10n --clean`.
+    - Add `cargo zng l10n --clean-deps` to remove previously copied localization before new copy.
+    - Add `cargo zng l10n --clean-template` to remove previously scraped files.
 * Add `cargo zng l10n --no-pkg` to skip scraping the target package, only copy localization from dependencies.
 * Don't show keyboard shortcuts in mobile menus.
 * Fix incorrect `TouchInputArgs::position` in nested windows.